+++ src/main/java/dev/fulmineo/companion_bats/entity/ai/goal/CompanionBatRangedAttackGoal.java
@@ -1,6 +1,7 @@
 package dev.fulmineo.companion_bats.entity.ai.goal;
 
 import dev.fulmineo.companion_bats.entity.CompanionBatEntity;
+import net.minecraft.entity.Entity;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.ai.goal.Goal;
 import net.minecraft.entity.ai.pathing.EntityNavigation;
@@ -8,24 +9,22 @@ import net.minecraft.entity.ai.pathing.EntityNavigation;
 public class CompanionBatRangedAttackGoal extends Goal {
 	private CompanionBatEntity entity;
 	private EntityNavigation navigation;
-	private int canStartCountdownTicks;
 	private int updateCountdownTicks;
 	private double minDistanceSquared;
 	private double maxDistanceSquared;
-	private int rangedAttackCooldown;
+	private RangedAttackFunction attackFunction;
 
-	public CompanionBatRangedAttackGoal(CompanionBatEntity entity, float minDistance, float maxDistance, int rangedAttackCooldown) {
+	public CompanionBatRangedAttackGoal(CompanionBatEntity entity, float minDistance, float maxDistance, RangedAttackFunction attackFunction) {
 		this.entity = entity;
         this.minDistanceSquared = (double)(minDistance * minDistance);
         this.maxDistanceSquared = (double)(maxDistance * maxDistance);
-		this.rangedAttackCooldown = rangedAttackCooldown;
-		this.canStartCountdownTicks = rangedAttackCooldown;
+		this.attackFunction = attackFunction;
         this.navigation = entity.getNavigation();
 	}
 
 	public boolean canStart() {
-        if (--this.canStartCountdownTicks <= 0) {
-			this.canStartCountdownTicks = 10;
+        if (--this.entity.rangedAttackTicks <= 0) {
+			this.entity.rangedAttackTicks = 10;
 			LivingEntity livingEntity = this.entity.getOwner();
 			if (livingEntity == null || livingEntity.isSpectator() || this.entity.isRoosting()) {
 				return false;
@@ -37,7 +36,7 @@ public class CompanionBatRangedAttackGoal extends Goal {
     }
 
     public boolean shouldContinue() {
-		return this.canStartCountdownTicks <= 0 && !this.navigation.isIdle() && this.entity.getTarget() != null && this.entity.getTarget().isAlive() && this.isWithinDistanceToAttack(this.entity.getTarget());
+		return this.entity.rangedAttackTicks <= 0 && !this.navigation.isIdle() && this.entity.getTarget() != null && this.entity.getTarget().isAlive() && this.isWithinDistanceToAttack(this.entity.getTarget());
     }
 
     public void start() {
@@ -52,8 +51,8 @@ public class CompanionBatRangedAttackGoal extends Goal {
 			this.updateCountdownTicks = 10;
             if (!this.entity.isLeashed() && !this.entity.hasVehicle()) {
 				LivingEntity target = this.entity.getTarget();
-				if (target != null && target.isAlive() && this.entity.getVisibilityCache().canSee(target) && this.entity.tryRangedAttack(target)){
-					this.canStartCountdownTicks = this.rangedAttackCooldown;
+				if (target != null && target.isAlive() && this.entity.getVisibilityCache().canSee(target)){
+					this.attackFunction.rangedAttack(target);
 				}
             }
         }
@@ -63,4 +62,9 @@ public class CompanionBatRangedAttackGoal extends Goal {
 		double distance = this.entity.squaredDistanceTo(target);
 		return distance > this.minDistanceSquared && distance < this.maxDistanceSquared;
 	}
+
+	public interface RangedAttackFunction {
+		void rangedAttack(Entity target);
+	}
 }
+